Android: moves android_webview/ to use long for JNI.
BUG=317523

Review URL: https://codereview.chromium.org/83203002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@237788 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: f454c268198472d6d4630ae2c19abb7b11d4b7e4

+++ android_webview/java/src/org/chromium/android_webview/AwAutofillManagerDelegate.java
@@ -19,17 +19,17 @@ import org.chromium.ui.autofill.AutofillSuggestion;
 @JNINamespace("android_webview")
 public class AwAutofillManagerDelegate {
 
-    private final int mNativeAwAutofillManagerDelegate;
+    private final long mNativeAwAutofillManagerDelegate;
     private AutofillPopup mAutofillPopup;
     private ViewGroup mContainerView;
     private ContentViewCore mContentViewCore;
 
     @CalledByNative
-    public static AwAutofillManagerDelegate create(int nativeDelegate) {
+    public static AwAutofillManagerDelegate create(long nativeDelegate) {
         return new AwAutofillManagerDelegate(nativeDelegate);
     }
 
-    private AwAutofillManagerDelegate(int nativeAwAutofillManagerDelegate) {
+    private AwAutofillManagerDelegate(long nativeAwAutofillManagerDelegate) {
         mNativeAwAutofillManagerDelegate = nativeAwAutofillManagerDelegate;
     }
 
@@ -85,5 +85,6 @@ public class AwAutofillManagerDelegate {
         array[index] = new AutofillSuggestion(name, label, uniqueId);
     }
 
-    private native void nativeSuggestionSelected(int nativeAwAutofillManagerDelegate, int position);
+    private native void nativeSuggestionSelected(long nativeAwAutofillManagerDelegate,
+            int position);
 }
