+++ android_webview/java/src/org/chromium/android_webview/AwSettings.java
@@ -114,7 +114,7 @@ public class AwSettings {
     private static boolean sAppCachePathIsSet = false;
 
     // The native side of this object. It's lifetime is bounded by the WebContent it is attached to.
-    private int mNativeAwSettings = 0;
+    private long mNativeAwSettings = 0;
 
     // A flag to avoid sending superfluous synchronization messages.
     private boolean mIsUpdateWebkitPrefsMessagePending = false;
@@ -224,7 +224,7 @@ public class AwSettings {
     }
 
     @CalledByNative
-    private void nativeAwSettingsGone(int nativeAwSettings) {
+    private void nativeAwSettingsGone(long nativeAwSettings) {
         assert mNativeAwSettings != 0 && mNativeAwSettings == nativeAwSettings;
         mNativeAwSettings = 0;
     }
@@ -1418,21 +1418,21 @@ public class AwSettings {
         }
     }
 
-    private native int nativeInit(int webContentsPtr);
+    private native long nativeInit(long webContentsPtr);
 
-    private native void nativeDestroy(int nativeAwSettings);
+    private native void nativeDestroy(long nativeAwSettings);
 
-    private native void nativeResetScrollAndScaleState(int nativeAwSettings);
+    private native void nativeResetScrollAndScaleState(long nativeAwSettings);
 
-    private native void nativeUpdateEverythingLocked(int nativeAwSettings);
+    private native void nativeUpdateEverythingLocked(long nativeAwSettings);
 
-    private native void nativeUpdateInitialPageScaleLocked(int nativeAwSettings);
+    private native void nativeUpdateInitialPageScaleLocked(long nativeAwSettings);
 
-    private native void nativeUpdateUserAgentLocked(int nativeAwSettings);
+    private native void nativeUpdateUserAgentLocked(long nativeAwSettings);
 
-    private native void nativeUpdateWebkitPreferencesLocked(int nativeAwSettings);
+    private native void nativeUpdateWebkitPreferencesLocked(long nativeAwSettings);
 
     private static native String nativeGetDefaultUserAgent();
 
-    private native void nativeUpdateFormDataPreferencesLocked(int nativeAwSettings);
+    private native void nativeUpdateFormDataPreferencesLocked(long nativeAwSettings);
 }
